Improved how URLs are constructed.
The frontend detects its own base URL from location.href and constructs URLs to itself using that. URLs to BlackLab Server are all constructed based on the blsUrlInternal and -External settings in the WhitelabBundle*.properties files.

Additionally, a divide by zero bug was fixed and the Home page text was incorporated into the .war for now because of http/https problems.